[PATCH] Example PyInstaller configuration
